api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pillow
requests
scikit-learn
numpy
python-multipart
